Fix date range for professional status emails job (#2339)
The nightly job that sends emails for QTS/EYTS awardees (and QTLS
losers) looks for events that have occurred since the last run. It does
this by consulting the `job_metadata` table and updating that same table
when the job has completed. But, the metadata is serialized as JSON, so
when a `DateTime` goes in it comes out as a `string`. This meant our `is
DateTime` check failed and we were falling back to the base DateTime in
configuration (which is circa 2023). As such, we're repeatedly sending
the same emails every single day.

I've updated the `JobMetadata.Metadata` type to be `Dictionary<string,
string>` so we avoid this non-obvious serialization behaviour and fixed
the type check in the job.

I've also moved the `TrsDbContext` creation to after the
`TransactionScope` to ensure it's created within the transaction. (I'm
not convinced this was working 100% correctly before.).