Stub out the Unlock teacher endpoint (#2350)
Now that Teacher Self Service portal is gone, unlocking a record makes
no sense. Find a lost TRN is the only service that calls this, and it
will shortly be changed to remove the call entirely. For now, we can
return a stub response and once Find has stopped calling the endpoint we
can remove it completely.