Upgrade XUnit packages (#2359)
Upgrades XUnit packages to take advantage of bug fixes including fixes
to the analyzers.

Upgrades the following packages:
* xunit to 2.9.3
* xunit.analyzers to1.23.0
* xunit.assert to 2.9.3
* xunit.runner.visualstudio to 3.1.3

Note: xunit.analyzers now added as a Top-level dependency whereas
previously it was a transitive dependency. This is to take advantage of
analyzer fixes such as https://github.com/xunit/xunit/issues/2852 (fixed
in 1.9.0)