Rewired Index page for change requests to get data from TRS support tasks instead of DQT (#2371)
### Context

The TRS Console currently allows users to manage Change name and change
date of birth requests which are stored as “Cases” in DQT.
These cases/requests are created when a teacher requests a change of
details via AYTQ.
As part of migrating from DQT, we need to move these to be stored as
support tasks in TRS.

### Changes proposed in this pull request

Re-wire the Index page at
`TeachingRecordSystem.SupportUi.Pages.ChangeRequests.EditChangeRequest`
to get its data from TRS i.e. `SupportTask` with `SupportTaskType` of
`SupportTaskType.ChangeNameRequest` or
`SupportTaskType.ChangeDateOfBirthRequest` + amend tests as appropriate.