Rewire reject page for name and DOB change requests to use TRS support tasks (#2377)
### Context

The TRS Console currently allows users to manage Change name and change
date of birth requests which are stored as “Cases” in DQT.
These cases/requests are created when a teacher requests a change of
details via AYTQ.
As part of migrating from DQT, we need to move these to be stored as
support tasks in TRS.

### Changes proposed in this pull request

Re-wire the `Reject` page at
`TeachingRecordSystem.SupportUi.Pages.ChangeRequests.EditChangeRequest`
to reject or cancel a change name or date of birth request
`SupportTask`.
Considerations:
- When the task is Rejected we need to replicate sending an email like
the `CaseRejectedSendEmailPlugin` does currently.
- Use `INotificationSender` to send emails via Notify in TRS (see the
API endpoints for prior art).
- Need to create a `ChangeNameRequestSupportTaskRejected` or
`ChangeNameRequestSupportTaskCancelled` event.