Add tile for change requests and amend support tasks index to get change requests from TRS (#2384)
### Context

The TRS Console currently allows users to manage Change name and change
date of birth requests which are stored as “Cases” in DQT.
These cases/requests are created when a teacher requests a change of
details via AYTQ.
As part of migrating from DQT, we need to move these to be stored as
support tasks in TRS.

### Changes proposed in this pull request

Amend the `Pages/SupportTasks/Index` page to get change date of birth
and name requests from TRS instead of DQT i.e. `SupportTask` with
`SupportTaskType` of `SupportTaskType.ChangeNameRequest` or
`SupportTaskType.ChangeDateOfBirthRequest` with Status
`SupportTaskStatus.Open`
Amend the `Pages/Index` page to add a tile for change requests which
links to the above page (i.e. `/support-tasks`).