Amended to use email templates from correct notify service (#2406)
### Context

The email templates for change name and date of birth requests are in
the Teaching Regulation Agency GOV Notify service whereas the email
templates for qualification emails in TRS are in the Teaching Regulation
service.

Now we’ve ported the change name and date of birth requests to support
tasks in TRS we need to be able to send the emails from the Teaching
Regulation service in GOV Notify.