Ensure DB transactions (#2388)
We have a few places where we take locks `select ... for update` and
where we schedule jobs where we require a transaction. This fixes both
of those places with two different approaches:

An extension method is added on `HttpContext` -
`EnsureDbTransactionAsync` - that checks the current request's
`TrsDbContext` and starts a transaction if required. A middleware will
commit that transaction when the request is done. This ensures that the
`select ... for update`s we do in filters keep the row locked for the
duration of the request.

For enqueueing Hangfire jobs, we need an ambient `TransactionScope`
instead. This adds a decorator to `IBackgroundJobScheduler` that checks
if there is a current `Transaction`, and throws if there is not. Our
existing `TransactionScopeAttribute` will deal with creating the
`TransactionScope`.