Ported TRN generation from external API to internal TRS service (#2422)
### Context

Now that DQT is (almost) gone, we can move TRN generation into TRS
itself.

### Changes proposed in this pull request

Copy the TRN generation logic from the TRN Generation API into TRS
itself.
Create an implementation of ITrnGenerator that uses TRS’ database (be
mindful of DB transactions and ensuring the allocation occurs in the
same transaction as other DB changes).

NOTE: the change to register the `DbTrnGenerator` implementation of
`ITrnGenerator` and ensure any calls are part of a transaction will be
in a separate PR