Add mechanism for constructing FormFlow state instances (#2467)
Currently we use a pattern where the first page of a journey calls an
`EnsureInitialized` method on journey state types to set up the initial
journey state. This is a little clunky and in some cases this method can
be called multiple times for a particular journey instance. It also
means we can't use `required` properties and everything has to be
nullable.

This adds an alternative mechanism - `IJourneyStateFactory<T>`. When the
journey is created, if a class is found implementing that interface then
its `CreateAsync()` method is called to create the state instance. We
pass along the `JourneyDescriptor` and `HttpContext` so that the factory
can access any identifiers for the journey instance.