Prevent users submitting NPQ TRN request multiple times (#2489)
We're getting some duplicate NPQ TRN requests coming through, likely
because users are hitting Submit multiple times.

This adds a 'select ... for update` on the journey state so that only
one request can be updating it at once. That combined with the flag we
set - `HasPendingTrnRequest` - when the request is actually created
should fix this.