Add SupportTaskTypeId to SupportTask (#2523)
We need to get our reference data into the DB so it can be reported on,
specifically the enum-based reference data like `SupportTaskType`.

This adds `SupportTaskTypeId` to `SupportTask` so we'll have both the
new and old values being written (`SupportTask.SupportTaskType` will be
replaced by `SupportTask.SupportTaskTypeId` eventually). A follow-up PR
will back-fill `SupportTaskTypeId` and remove usage of
`SupportTaskType`.