Make transaction handling consistent (#2534)
Currently we have a mish-mash of approaches for handling transactions.
In some places we do `DbContext.Database.BeginTransaction`, in some
places we use `new TransactionScope...`. Sometimes transactions are
created by filters or middleware, sometimes explicitly in the page or
API handler.

This adds middleware to all our web apps that creates a
`TransactionScope` then commits it when the request is done. With that
we can remove all other transaction management (in the web apps at
least).

I've had to use `TransactionScope` instead of
`DbContext.Database.BeginTransaction` as it's the only way to work with
Hangfire's storage implementation. It's also a little more robust and
should work even if we end up creating multiple DbContexts for a
request.