Validate out when creating a new person that has an empty first name or lastname (#2560)
Capita files sometimes do not have a firstname or middle name, however
there columns in the db are non nullable. This updates the import to use
an empty string if creating a new record and they are not present.

- Update import so that if the result of importing a row is to create a
person, make sure that the firstname and last name are present

### Checklist

-   [] Attach to Trello card
-   [x] Rebased master
-   [x] Cleaned commit history
-   [x] Tested by running locally
-   [x] Run DQT integration tests locally (if appropriate)