Add names and national_insurance_numbers columns to person (#2563)
We're moving away from `person_search_attributes` for search towards
using `persons` directly (with a GIN index). For that we need to include
the names from `previous_names` and the NINOs from `tps_employments`.
This adds three new array fields - `names`, `last_names` and
`national_insurance_numbers` with the appropriate triggers to populate
them.

A follow up PR will back-fill, then we can create the index and start
using it.