Add SupportTaskTypes to the DB (#2575)
D&I need to report on support task progress and so the tasks need to be
pushed to the reporting DB. This adds task types to the TRS DB and
alters the publication to sync it to the reporting DB. Tasks themselves
will follow later.

This removes the previously-added `SupportTaskTypeId`; the model I had
in mind with that didn't work well with EF so I abandoned it.