Write support task type into data payload (#2579)
We have a weird setup with the `Data` property on `SupportTask`
currently where we have to explicitly JSON (de|)serialize. The first
step of moving to a simpler setup is to include the support task type in
the data payload so the serializer knows which concrete type to
deserialize to.

This amends the serialization process to add that property. A follow up
PR will back-fill then we remove the explicit serialization completely.