prevent importing capita trn that already exists but is de-activated (#2602)
Update capita import to error if it finds an existing trn for the row
that is being imported, but is de-activated.

**context:**
We received a record through the capita interface that would of resulted
in new person being created with a trn that had been long de-activated.
This PR updates the validation to fetch de-activated records and return
an error if the record is de-activated.