Added missing Index pages and corrected "resolve" paths (#2612)
### Context

A number of journeys in the TRS Console are missing an Index page + some
of the URLs are not quite right e.g. missing /resolve for the resolve
support task journeys.

### Changes proposed in this pull request

Add an Index page for any journeys which are missing one (this can
activate the journey and simply redirect to the first “proper” page in
the journey) OR rename the obvious first page in the journey to be
Index.
Correct paths as necessary.