Refactored link generator (#2639)
A single helper class `TrsLinkGenerator` is used to generate the URL to
navigate between pages in the TRS Support Console.
It’s become unwieldy now we have so many UI journeys and there are
naming inconsistencies.

This replaces `TrsLinkGenerator` with `SupportUiLinkGenerator` which is
composed of nested link generator classes specific to each journey.
The code to get a link now looks like this sort of thing: 
`linkGenerator.RoutesToProfessionalStatus.EditRoute.CheckAnswers(...)`