Use xUnit v3 for Cli.Tests (#2685)
This moves Cli.Tests to xUnit v3. There's no DI library in here as the
next xUnit release will include a mechanism for controlling how types
are created so we can use that when it lands.

I've also had to separate library projects for xUnit v3 to avoid clashes
with v2. Once everything is migrated these will be consolidated.

https://trello.com/c/i15vfgVZ