point trs integration jobs to sftp storage container (#2643)
https://trello.com/c/CMuoKrz8/1719-replace-globalscape-sftp-with-sftp-enabled-storage-container

Sftp has been enabled on a new storage account, previously the storage
account received it's files that were uploaded to a dqt sftp server, and
moved them to a new folder once processed by trs. We are moving away
from a self hosted sftp server, meaning that we need to add sftp users
to an sftp enabled storage account. Because of this we can no longer
just move them to a new folder in the storage account, as they can be
deleted or modified.

- re-wire ewc/capita jobs so that it reads and writes to the new sftp
enabled storage container.
- move processed files to another storage container that sftp users
cannot modify.
- sftp enabled storage accounts cannot be accessed using the standard
**BlobServiceClient**. This has been replaced with
**DataLakeServiceClient**