Added backfill job to get support tasks data in reporting db (#2760)
### Context

D&I need to access to the data in the `support_tasks` table for
reporting purposes.

### Changes proposed in this pull request

Add a job that back-fills `support_tasks` into the reporting DB (it is
already in the database publication but that was only added recently so
there will be missing data in the reporting db).