Move Core.Tests into Core.LegacyTests (#2764)
`Core.Tests` is now the only project not migrated to xUnit v3. It's the
most complex to move since it has lots of custom fixtures. This renames
`Core.Tests` to `Core.LegacyTests` and adds a new `Core.Tests` that uses
xUnit v3. We can gradually migrate tests to the new project one-by-one.