Migrate Add note process to new event model (#2826)
We're about to create audit events for DQT-created notes before we drop
DQT altogether. This migrates the current Add Note process in TRS to use
the new event model so that all note-related events are using the new
model. I've also moved the note and event creation into a Service.

A follow up PR will back-fill events for notes created in TRS.