Add a means to track journey steps to FormFlow (#2849)
Currently, there are a lot of chores to do when implementing a multi
page journey. This PR attempts to generalise some of the patterns we use
so that these chores can be handled by more general purpose mechanisms.
It does this by adding an `IJourneyWithSteps` interface and a
`JourneySteps` type and a host of helper methods for working with them.

`JourneySteps` is essentially the list of pages that are a valid for a
given journey instance for a point in time. When a step is completed, it
pushes the next step into the list of steps. At that point, the
preceding and/or proceeding steps can optionally be removed. (This is
primarily for when a user goes back and changes a decision that
invalidates previously valid steps. It's also useful for if there are
steps in a journey that we don't want a user to be able to go back to
after a particular time.)

With this, we can easily check if a particular page is valid for a given
journey (by checking if the step exists in the list of steps) and get
the previous page (by looking at the preceding step in the list of
steps).

A few things internal to FormFlow had to change to make this happen.
- The state creation delegates are now passed the instance ID. This is
essential as state creation delegates need to be able to know the
initial page for a journey (which requires the instance ID).
- The `properties` collection on `JourneyInstance` is no-longer
persisted. I need to use this to stash the `ActionContext` on and we've
never used it for persisting anything with the instance any way.

A follow-up PR will show this in use for the Create Alert journey.