Fix bugs in Capita import, refactor tests (#2845)
Fixes some inconsistencies in Capita import, also refactors/rewrites
tests as some of them were dependent on a method in the job that needed
to be removed. Tests were rewritten to be (hopefully) easier to
understand (e.g. factored out CSV and assertion logic) and corrected
test names based on actual behaviour.

https://trello.com/c/VQbisG3I/1967-fix-capita-import-job