Add Create method to SupportTaskService (#2868)
We're moving to DB-generated IDs for support tasks. The older style
event generation doesn't work with DB-generated data as the event is
created before the row is saved to the DB. As such, we need to move
support task generation to the newer style events. This is the first
step - adding a `CreateSupportTaskAsync` method to the
`SupportTaskService`.

I've also added tests for the `Delete` method.