Remove ApplicationUser ShortName (#2880)
### Context

We are using many tag across the service, and are not always consistent
with how these are being used. We need to review where tags are being
used, and set out a definitive approach to how these are being used.


https://trello.com/c/amTAznUI/1992-update-tags-across-the-service-so-that-they-are-consistent

### Changes proposed in this pull request

Removes ShortName property from ApplicationUser - we will use the full
ApplicationUser name rather than the short name, and just shorten the
full name if needed.

### Checklist

-   [x] Attach to Trello card
-   [x] Rebased master
-   [x] Cleaned commit history
-   [x] Tested by running locally