Fix tag inconsistencies (#2874)
### Context

We are using many tag across the service, and are not always consistent
with how these are being used. We need to review where tags are being
used, and set out a definitive approach to how these are being used.


https://trello.com/c/amTAznUI/1992-update-tags-across-the-service-so-that-they-are-consistent

### Changes proposed in this pull request

* Introduces `<tag>` TagHelper which takes a `TagColor` enum value so
all the colour options are visible in autocomplete when adding/editing a
tag and subsumes `<colored-tag>` functionality
* Introduces `TagColor` enum containing
* a value for each of the allowed colours for tags [as defined
here](https://design-system.service.gov.uk/components/tag/#additional-colours)
* `Auto` which generates a colour based on the tag content - what used
to be `<coloured-tag>`

### Checklist

-   [x] Attach to Trello card
-   [x] Rebased master
-   [x] Cleaned commit history
-   [x] Tested by running locally