Move TRN request logic into service (#2886)
This pulls most of the logic around TRN requests into the
`TrnRequestService` and adds tests for all the methods on that service.
Follow up PRs will re-wire the various journeys to use this and as such
there are a few `TODO`s still about. The matching logic should probably
also move here too but I'll address that separately.