Handle invalid email addresses when calling TRN token API (#2902)
We've had some invalid email addresses creep in with TRN requests. This
updates the API call to the TRN token API to check that the email
address is valid before calling it.

I've also updated the TRN request API's validation to use our rules for
email addresses rather than FV's.