Handle infinity and -infinity in reporting sync (#2952)
Some of the entries in `events` have a `-infinity` `inserted` value,
which blows up when syncing to the reporting DB due to being out of
range. This amends the service to convert those values to `null`. (This
column is largely redundant so its value in the reporting DB isn't super
important.)